# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_kiarashAlizadeh_herotech_internal_dto.AuctionResponse:
    properties:
      ends_at:
        type: string
      highest_bid:
        type: integer
      id:
        type: string
      item_id:
        type: string
      seller_id:
        type: string
      start_price:
        type: integer
      status:
        type: string
      winner_id:
        type: string
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.CreateAuctionRequest:
    properties:
      duration:
        description: Duration in hours
        type: integer
      item_id:
        type: string
      start_price:
        type: integer
    type: object
...
  ? github_com_kiarashAlizadeh_herotech_internal_dto.PaginatedResponse-github_com_kiarashAlizadeh_herotech_internal_dto_AuctionResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.AuctionResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PaginationMeta'
    type: object
...
  github_com_kiarashAlizadeh_herotech_internal_dto.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.PlaceBidRequest:
    properties:
      amount:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@Aethoria.com
    name: API Support
  description: Core infrastructure backend for managing Guild Wallets, Item Trading,
    and Legendary Auctions.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DRAGON MARKET PLACE API
  version: "1.0"
paths:
  /auctions:
    get:
      description: Streams current dynamic storefront auction indices filtering active
        runs
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PaginatedResponse-github_com_kiarashAlizadeh_herotech_internal_dto_AuctionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List active marketplace auctions
      tags:
      - auctions
    post:
      consumes:
      - application/json
      description: Commences a public competitive auction profile cycle sequence restricted
        exclusively to legendary assets
      parameters:
      - description: Seller Guild UUID
        in: header
        name: X-Guild-ID
        required: true
        type: string
      - description: Auction listing duration configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.CreateAuctionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.AuctionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Launch a Legendary item auction
      tags:
      - auctions
  /auctions/{id}:
    get:
      description: Retrieves current auction configurations, ends timetables, and
        leading winner parameters
      parameters:
      - description: Auction UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.AuctionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get auction board details
      tags:
      - auctions
  /auctions/{id}/bid:
    delete:
      description: Cancels bid stakes and unlocks reserved allocations instantly if
        bidder isn't holding the pole position
      parameters:
      - description: Bidder Guild UUID
        in: header
        name: X-Guild-ID
        required: true
        type: string
      - description: Auction UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retract a non-leading bid position
      tags:
      - auctions
    post:
      consumes:
      - application/json
      description: Evaluates and appends a bid reservation position. Requires at least
        5% gain increment steps.
      parameters:
      - description: Bidder Guild UUID
        in: header
        name: X-Guild-ID
        required: true
        type: string
      - description: Auction UUID
        in: path
        name: id
        required: true
        type: string
      - description: Bid financial allocation criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PlaceBidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a new competitive bid
      tags:
      - auctions
...
swagger: "2.0"